# perf: improve copyZeroAlloc for os.File and net.TCPConn